custom_components/requirements_dev.txt
ruff>=0.9.0
mypy>=1.14.0
homeassistant-stubs
